apis/audio/pyproject.toml
[tool.poetry]
name = "que-vino-audio-api"
version = "1.0.0"
description = "Microservicio de generación de audio (TTS) para Que Vino!?"
authors = ["Antigravity <antigravity@google.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.2.0"
google-genai = ">=1.70.0"
google-cloud-storage = "^2.14.0"
google-cloud-bigquery = "^3.18.0"
google-cloud-texttospeech = "^2.16.0"
httpx = "^0.28.1"
python-json-logger = "^2.0.7"
cachetools = "^5.3.3"
firebase-admin = "^6.4.0"
python-multipart = "^0.0.9"
mcp = {extras = ["cli"], version = "^1.2.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]
